requirements.txt
Flask-SocketIO==0.6.0
future==0.14.1
elasticsearch==1.6.0
arrow==0.6.0
gunicorn==19.3.0